Time: 2 ms (56.30%) | Memory: 7.9 MB (41.35%) - LeetSync

+++ 278-first-bad-version/first-bad-version.cpp
@@ -4,7 +4,7 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        int low = 1, high = n, mid;
+        long long int low = 1, high = n, mid;
         while (low < high) {
             mid = low + (high - low) / 2;
             if (isBadVersion(mid)) {
